casing error solved, renaming context file back to InteleosContext.tsx
